--- ananke/envs.py
@@ -1,5 +1,0 @@
-
-EBF_BASEDIR = "/scratch/05328/tg846280/FIRE_Public_Simulations/gaia_mocks_ebf"
-HDF5_BASEDIR = "/scratch/05328/tg846280/FIRE_Public_Simulations/gaia_mocks_hdf5"
-DR3_PRESF_BASEDIR = "/scratch/05328/tg846280/FIRE_Public_Simulations/ananke_dr3/preSF"
-DR3_BASEDIR = "/scratch/05328/tg846280/FIRE_Public_Simulations/ananke_dr3"
