+++ pyproject.toml
@@ -1,0 +1,20 @@
+[build-system]
+requires = ["setuptools", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.poetry]
+name = "ananke"
+version = "1.0.0"
+description = "Converting Latte mock catalog into the Ananke DR3 synthetic surveys"
+authors = ["Tri Nguyen <tnguy@mit.edu>", "Xiaowei Ou <xwou@mit.edu>"]
+license = "MIT"
+readme = "README.md"
+
+[tool.poetry.license]
+file = "LICENSE"
+
+[tool.poetry.dependencies]
+python = "^3.9"
+numpy = "^1.21.0"
+scipy = "^1.7.0"
+pandas = "^1.3.0"
