--- requirements.txt
@@ -1,9 +1,0 @@
-numpy
-astropy
-scipy
-pandas
-h5py
-pygaia
-tqdm
-numba
-git+https://github.com/sanjibs/ebfpy.git@master
