+++ scripts/calc_props.py
@@ -1,13 +1,12 @@
 #!/usr/bin/env python
 
-import os
-import sys
-import h5py
 import argparse
-import logging
+import h5py
+import os
 import time
 
-from .. import coordinates, errors, extinction, io, flags, envs
+from ananke import coordinates, errors, extinction, io, flags, config
+from ananke.logger import logger
 
 def parse_cmd():
     parser = argparse.ArgumentParser()
@@ -30,31 +29,20 @@ def parse_cmd():
                         help='Batch size')
     return parser.parse_args()
 
-def set_logger():
-    logger = logging.getLogger(__name__)
-    logger.setLevel(logging.DEBUG)
-    formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(message)s')
-    stream_handler = logging.StreamHandler(sys.stdout)
-    stream_handler.setFormatter(formatter)
-    logger.addHandler(stream_handler)
-    return logger
-
-def main(FLAGS, LOGGER=None):
+def main(FLAGS):
     """ Calculate catalog properties """
     gal = FLAGS.gal
     lsr = FLAGS.lsr
     rslice = FLAGS.rslice
-    if LOGGER is None:
-        LOGGER = set_logger()
 
     # get file information from galaxy, lsr, and rslice
     in_path = os.path.join(
-        envs.DR3_PRESF_BASEDIR, f"{gal}/lsr-{lsr}",
+        config.DR3_PRESF_BASEDIR, f"{gal}/lsr-{lsr}",
         f"lsr-{lsr}-rslice-{rslice}.{gal}-res7100-md-sliced-gcat-dr3.{FLAGS.ijob}.hdf5")
 
     # Read in and calculate extinction magnitude
-    LOGGER.info("Calculate extra coordinates, extincted magnitudes, and errors")
-    LOGGER.info(f"In: {in_path}")
+    logger.info("Calculate extra coordinates, extincted magnitudes, and errors")
+    logger.info(f"In: {in_path}")
 
     with h5py.File(in_path, 'a') as f:
         # Add header
@@ -68,7 +56,7 @@ def main(FLAGS, LOGGER=None):
         N_batch = (N + FLAGS.batch_size - 1) // FLAGS.batch_size
 
         for i_batch in range(N_batch):
-            LOGGER.info(f'Progress [{i_batch}/{N_batch}]')
+            logger.info(f'Progress [{i_batch}/{N_batch}]')
             i_start = i_batch * FLAGS.batch_size
             i_stop = i_start + FLAGS.batch_size
             indices = (i_start, i_stop)
@@ -100,13 +88,12 @@ def main(FLAGS, LOGGER=None):
 
 if __name__ == "__main__":
     FLAGS = parse_cmd()
-    LOGGER = set_logger()
 
     # run main and keep track of time
     t0 = time.time()
-    main(FLAGS, LOGGER)
+    main(FLAGS)
     t1 = time.time()
 
-    LOGGER.info(f"Total run time: {t1 - t0}")
-    LOGGER.info("Done!")
+    logger.info(f"Total run time: {t1 - t0}")
+    logger.info("Done!")
