+++ scripts/rotate_coords.py
@@ -1,18 +1,16 @@
 #!/usr/bin/env python
 
-import os
-import sys
-import h5py
 import argparse
-import logging
+import h5py
+import os
 import time
 
 import numpy as np
-import astropy
 import astropy.coordinates as coord
 import astropy.units as u
 
-from .. import photometric_utils, io, envs
+from ananke import photometric_utils, io, config
+from ananke.logger import logger
 
 FLAGS = None
 
@@ -26,15 +24,6 @@ def parse_cmd():
     parser.add_argument('--batch-size', required=False, type=int, default=1000000,
                         help='Batch size')
     return parser.parse_args()
-
-def set_logger():
-    logger = logging.getLogger(__name__)
-    logger.setLevel(logging.DEBUG)
-    formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(message)s')
-    stream_handler = logging.StreamHandler(sys.stdout)
-    stream_handler.setFormatter(formatter)
-    logger.addHandler(stream_handler)
-    return logger
 
 def rotate_coords_ananke(x, lsr):
     """ Rotate coordinate based on LSR """
@@ -92,29 +81,27 @@ def calc_new_coords(data, lsr, indices=(None, None)):
 
     return new_data
 
-def main(FLAGS, LOGGER=None):
+def main(FLAGS):
     """ Rotate Ananke coordinate """
     gal = FLAGS.gal
     lsr = FLAGS.lsr
     rslice = FLAGS.rslice
-    if LOGGER is None:
-        LOGGER = set_logger()
 
     # get file information from galaxy, lsr, and rslice
     in_path = os.path.join(
-        envs.DR3_PRESF_BASEDIR, f"{gal}/lsr-{lsr}",
+        config.DR3_PRESF_BASEDIR, f"{gal}/lsr-{lsr}",
         f"lsr-{lsr}-rslice-{rslice}.{gal}-res7100-md-sliced-gcat-dr3.{FLAGS.ijob}.hdf5")
 
     # Compute new astrometric coordinates
-    LOGGER.info("Rotate coordinate")
-    LOGGER.info(f"In  : {in_path}")
+    logger.info("Rotate coordinate")
+    logger.info(f"In  : {in_path}")
 
     with h5py.File(in_path, 'a') as f:
         N = len(f['dmod_true'])
         N_batch = (N + FLAGS.batch_size - 1) // FLAGS.batch_size
 
         for i_batch in range(N_batch):
-            LOGGER.info(f'Progress [{i_batch}/{N_batch}]')
+            logger.info(f'Progress [{i_batch}/{N_batch}]')
             i_start = i_batch * FLAGS.batch_size
             i_stop = i_start + FLAGS.batch_size
             indices = (i_start, i_stop)
@@ -124,13 +111,11 @@ def main(FLAGS, LOGGER=None):
 
 if __name__ == "__main__":
     FLAGS = parse_cmd()
-    LOGGER = set_logger()
 
     # run main and keep track of time
     t0 = time.time()
-    main(FLAGS, LOGGER)
+    main(FLAGS)
     t1 = time.time()
-
-    LOGGER.info(f"Total run time: {t1 - t0}")
-    LOGGER.info("Done!")
+    logger.info(f"Total run time: {t1 - t0}")
+    logger.info("Done!")
