+++ setup.cfg
@@ -1,0 +1,27 @@
+[metadata]
+name = ananke-dr3
+version = 1.0.0
+author = Tri Nguyen
+author_email = tnguy@mit.edu
+description = Converting Latte mock catalog into the Ananke DR3 synthetic surveys
+url = https://github.com/trivnguyen/ananke_dr3/
+license = MIT
+license_file = LICENSE
+
+[options]
+packages = find:
+package_dir =
+    = src
+install_requires =
+    # List your requirements here
+    astropy
+    h5py
+    numba
+    numpy
+    pandas
+    pygaia
+    scipy
+    tqdm
+
+[options.packages.find]
+where = src
