+++ src/ananke/config.py
@@ -1,4 +1,9 @@
 
+EBF_BASEDIR = "/scratch/05328/tg846280/FIRE_Public_Simulations/gaia_mocks_ebf"
+HDF5_BASEDIR = "/scratch/05328/tg846280/FIRE_Public_Simulations/gaia_mocks_hdf5"
+DR3_PRESF_BASEDIR = "/scratch/05328/tg846280/FIRE_Public_Simulations/ananke_dr3/preSF"
+DR3_BASEDIR = "/scratch/05328/tg846280/FIRE_Public_Simulations/ananke_dr3"
+
 ALL_MOCK_KEYS = {
     'parentid': 'parentid',
     'partid': 'partid',
@@ -41,4 +46,3 @@ ALL_EXT_KEYS = {
     'a0': 'A0',
     'ebv': 'ebv',
 }
-
