+++ src/ananke/data/LogErrVsMagSpline.csv
