+++ src/ananke/errors/photometric.py
