+++ src/ananke/errors/spectroscopic.py
